노드 선택 알고리즘 구현현

+++ src/app/algorithm/node_select.py
@@ -4,7 +4,7 @@ def calc_node_completion_time(node):
     return node['ocr_task_size'] / node['ocr_perf'] + node['trans_task_size'] / node['trans_perf']
 
 
-def select_node_for_ocr(ocr_task_size: int):
+def select_node_for_ocr(ocr_task_size: float):
     nodes = get_available_nodes()
     if len(nodes) == 0:
         return None
@@ -13,7 +13,7 @@ def select_node_for_ocr(ocr_task_size: int):
     return nodes[0]
 
 
-def select_node_for_trans(trans_task_size: int):
+def select_node_for_trans(trans_task_size: float):
     nodes = get_available_nodes()
     if len(nodes) == 0:
         return None
